web/sqli+rce/requirements.txt
flask
pyjwt
python-dotenv